+++ plugins/analytics-module-matomo/config.d.ts
@@ -1,0 +1,19 @@
+export interface Config {
+  app?: {
+    analytics: {
+      matomo: {
+        /**
+         * Matomo host URL
+         * @visibility frontend
+         */
+        host: string;
+
+        /**
+         * Matomo siteId for the Backstage Website
+         * @visibility frontend
+         */
+        siteId: string;
+      };
+    };
+  };
+}