+++ plugins/analytics-module-matomo/dev/Playground.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const Playground = () => {
+  return (
+    <>
+      <a href="#clicked">Click Here</a>
+    </>
+  );
+};